Methodology
I used my notes from a Microsoft Course to begin this course project. Next step is to write in a clear mood how to do it of this things. Part two of the course focus on statiscs with excel and there will be a second course that will use the aproach of this one to get more out of excel using data model and dax language.